requirements.txt
psutil
# docker-py should ensure >= 1.8.1
docker-py==1.8.1
